# Reduce unnecessary queries if crash/error params are empty