* Add logging around analytics ingester calls to see the performance impact

Before:

```
Started POST "/did_launch" for ::1 at 2017-03-28 13:00:57 -0400
  ActiveRecord::SchemaMigration Load (0.6ms)  SELECT "schema_migrations".* FROM "schema_migrations"
Processing by BaconsController#did_launch as */*
  Parameters: {"versions"=>"{\"fastlane\":\"2.19.1\"}", "steps"=>"{\"fastlane\":1}", "error"=>"", "crash"=>"", "fastfile_id"=>""}
  Bacon Load (129.4ms)  SELECT  "bacons".* FROM "bacons" WHERE "bacons"."action_name" = $1 AND "bacons"."tool_version" = $2 AND "bacons"."launch_date" = '2017-03-28' LIMIT 1  [["action_name", "fastlane"], ["tool_version", "2.19.1"]]
   (0.1ms)  BEGIN
  SQL (1.6ms)  UPDATE "bacons" SET "launches" = $1, "updated_at" = $2 WHERE "bacons"."id" = $3  [["launches", 1085], ["updated_at", "2017-03-28 17:00:57.913384"], ["id", 538309]]
   (0.9ms)  COMMIT
  Bacon Load (125.4ms)  SELECT  "bacons".* FROM "bacons" WHERE "bacons"."action_name" = $1 AND "bacons"."launch_date" = $2 AND "bacons"."tool_version" = $3 LIMIT 1  [["action_name", ""], ["launch_date", "2017-03-28"], ["tool_version", "unknown"]]
  CACHE (0.0ms)  SELECT  "bacons".* FROM "bacons" WHERE "bacons"."action_name" = $1 AND "bacons"."launch_date" = $2 AND "bacons"."tool_version" = $3 LIMIT 1  [["action_name", ""], ["launch_date", Tue, 28 Mar 2017], ["tool_version", "unknown"]]
```

Note second **Bacon Load** for `action_name = ""` and subesquent identical query that winds up hitting the Rails internal **CACHE**

After:

```
Started POST "/did_launch" for ::1 at 2017-03-28 13:02:38 -0400
Processing by BaconsController#did_launch as */*
  Parameters: {"versions"=>"{\"fastlane\":\"2.19.1\"}", "steps"=>"{\"fastlane\":1}", "error"=>"", "crash"=>"", "fastfile_id"=>""}
  Bacon Load (3.1ms)  SELECT  "bacons".* FROM "bacons" WHERE "bacons"."action_name" = $1 AND "bacons"."tool_version" = $2 AND "bacons"."launch_date" = '2017-03-28' LIMIT 1  [["action_name", "fastlane"], ["tool_version", "2.19.1"]]
   (0.1ms)  BEGIN
  SQL (0.4ms)  UPDATE "bacons" SET "launches" = $1, "updated_at" = $2 WHERE "bacons"."id" = $3  [["launches", 1086], ["updated_at", "2017-03-28 17:02:38.820943"], ["id", 538309]]
   (0.5ms)  COMMIT
```

Note that the `action_name = ""` queries are eliminated